clean up main.c
