+++ src/hardware.h
@@ -100,6 +100,7 @@ uint8_t temperature_level(void);
 void vtx_paralized(void);
 
 void timer_task();
+void RF_Delay_Init();
 
 #if defined HDZERO_FREESTYLE || HDZERO_FREESTYLE_V2
 extern uint8_t powerLock;
